+++ Samples/NP.Common.AspNetCore.Sample/Pages/Index.cshtml.cs
@@ -3,18 +3,17 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace NP.Common.AspNetCore.Sample.Pages
 {
-    public class IndexModel : PageModel
-    {
-        private readonly ILogger<IndexModel> _logger;
+	public class IndexModel : PageModel
+	{
+		private readonly ILogger<IndexModel> _logger;
 
-        public IndexModel(ILogger<IndexModel> logger)
-        {
-            _logger = logger;
-        }
+		public IndexModel(ILogger<IndexModel> logger)
+		{
+			_logger = logger;
+		}
 
-        public void OnGet()
-        {
-
-        }
-    }
+		public void OnGet()
+		{
+		}
+	}
 }