+++ src/NP.Common/Extensions/ObjectExtensions.cs
@@ -1,4 +1,7 @@
-﻿namespace NP.Common.Extensions
+﻿using System.Collections.Generic;
+using System;
+
+namespace NP.Common.Extensions
 {
 	/// <summary>
 	/// 
@@ -13,7 +16,7 @@
 		/// <returns></returns>
 		public static TTo As<TTo>(this object obj)
 		{
-			if(obj is TTo y)
+			if (obj is TTo y)
 			{
 				return y;
 			}
@@ -26,6 +29,53 @@
 		/// </summary>
 		/// <param name="obj"></param>
 		/// <returns></returns>
-		public static bool IsNull(this object obj) => obj  == null;
+		public static bool IsNull(this object obj) => obj == null;
+
+		/// <summary>
+		/// Clones the <paramref name="source"/>.
+		/// </summary>
+		/// <typeparam name="T"></typeparam>
+		/// <param name="source"></param>
+		/// <returns>The cloned object.</returns>
+		public static T Clone<T>(this T source) where T : ICloneable
+		{
+			if (source == null)
+			{
+				return default(T);
+			}
+
+			return (T)source.Clone();
+		}
+
+		/// <summary>
+		/// Clones each item of <paramref name="sources"/>.
+		/// </summary>
+		/// <typeparam name="T"></typeparam>
+		/// <param name="sources"></param>
+		/// <returns>The collection of cloned objects.</returns>
+		public static IEnumerable<T> Clone<T>(this IEnumerable<T> sources) where T : ICloneable
+		{
+			if (sources == null)
+			{
+				yield break;
+			}
+
+			foreach (var source in sources)
+			{
+				yield return source.Clone<T>();
+			}
+		}
+
+		/// <summary>
+		/// Format <paramref name="stringToFormat"/> with <paramref name="args"/>.
+		/// </summary>
+		/// <param name="stringToFormat">The string to be formatted.</param>
+		/// <param name="args">The arguments that will be used to format the string.</param>
+		/// <returns>The formatted string.</returns>
+		public static string With(this string stringToFormat, params object[] args)
+		{
+			return string.Format(stringToFormat, args);
+		}
+
 	}
 }